+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
  - Add support for Symfony 5.0 (#266, thanks to @Big-Shark)
  - Drop support for Symfony < 3.4 (#277)
+ - Add default value for the `release` option, using the detected root package version (#291 #292, thanks to @Ocramius)
 
 ## 3.2.1 (2019-12-19)
  - Fix handling of command with no name on `ConsoleListener` (#261)